adding put and post methods to places, reviews, users, and amenities

+++ api/v1/views/amenities.py
@@ -1,42 +1,53 @@
 #!/usr/bin/python3
 """ amenities """
 from api.v1.views import app_views
-from flask import jsonify, Blueprint, make_response, abort
+from flask import jsonify, Blueprint, make_response, abort, request
 from models import storage
 from models.amenity import Amenity
 from models.base_model import BaseModel
 
 
-@app_views.route('/amenities', methods=["GET"], strict_slashes=False)
+@app_views.route('/amenities', methods=["GET", "POST"],
+                 strict_slashes=False)
 def get_all_amenities():
     """ retrieves all amenity objects """
     output = []
     amenities = storage.all(Amenity).values()
-    for amenity in amenities:
-        output.append(amenity.to_dict())
-    return (jsonify(output))
+    if request.method == "GET":
+        for amenity in amenities:
+            output.append(amenity.to_dict())
+        return (jsonify(output))
+    if request.method == "POST":
+        data = request.get_json()
+        if not request.is_json():
+            abort(400, description="Not a JSON")
+        if 'name' not in request.json:
+            abort(400, description="Missing name")
+        amenity = Amenity(**data)
+        amenity.save()
+        return (jsonify(amenity.to_dict()), 201)
 
 
-@app_views.route('/amenities/<amenity_id>', methods=["GET"],
-                 strict_slashes=False)
+@app_views.route('/amenities/<amenity_id>', methods=[
+                 "GET", "PUT", "DELETE"], strict_slashes=False)
 def get_an_amenity(amenity_id):
     """ retrieves one unique amenity object """
-    amenities = storage.all(Amenity).values()
-    for amenity in amenities:
-        if amenity.id == amenity_id:
-            output = amenity.to_dict()
-            return (jsonify(output))
-    abort(404)
-
-
-@app_views.route('/amenities/<amenity_id>', methods=["GET", "DELETE"],
-                 strict_slashes=False)
-def del_an_amenity(amenity_id):
-    """ delete one unique amenity object """
-    amenity = storage.get(Amenity, amenity_id)
+    amenitiy = storage.get(Amenity, amenity_id)
     if amenity is None:
         abort(404)
-    storage.delete(amenity)
-    storage.save()
-    result = make_response(jsonify({}), 200)
-    return result
+    if request.method == "GET":
+        output = amenity.to_dict()
+        return (jsonify(output))
+    if request.method == "PUT":
+        data = request.get_json()
+        if not request.is_json():
+            abort(400, description="Not a JSON")
+        for key, value in data.items():
+            setattr(amenity, key, value)
+        amenity.save()
+        return (jsonify(amenity.to_dict()), 200)
+    if request.method == "DELETE":
+        storage.delete(amenity)
+        storage.save()
+        result = make_response(jsonify({}), 200)
+        return result
